normalized some type insanity

+++ riscv/insns/custom3_rs1_rs2.h
@@ -13,12 +13,12 @@ MMU.load_slow_path(RS2, 4096, (uint8_t*)(candidates), 0);
 uint8_t cacheline[cacheline_size];
 memcpy(cacheline, p->pb.cacheline_words, cacheline_size);
 
-short mem_type = p->pb.mem_type;
+int mem_type = p->pb.mem_type;
 
 //Run external command
 std::string cmd;
 if (mem_type == 0) //data 
-    cmd = construct_sdecc_data_recovery_cmd(MMU.data_sdecc_script_filename, victim_message, candidates, cacheline, msg_size, cacheline_size/msg_size, (unsigned)(blockpos));
+    cmd = construct_sdecc_data_recovery_cmd(MMU.data_sdecc_script_filename, victim_message, candidates, cacheline, msg_size, cacheline_size/msg_size, blockpos);
 else //instruction
     cmd = construct_sdecc_inst_recovery_cmd(MMU.inst_sdecc_script_filename, victim_message, candidates, msg_size);
